+++ discO/plugin/gala/__init__.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+# see LICENSE.rst
+
+""":mod:`~gala` interface."""
+
+__all__ = [
+    "GalaPotentialWrapper",
+]
+
+
+##############################################################################
+# IMPORTS
+
+# PROJECT-SPECIFIC
+from .wrapper import GalaPotentialWrapper
+
+##############################################################################
+# END