+++ discO/plugin/gala/tests/__init__.py
@@ -1,11 +1,9 @@
 # -*- coding: utf-8 -*-
 # see LICENSE.rst
 
-"""Test :mod:`~discO.plugin.agama`."""
+"""Test :mod:`~discO.plugin.gala`."""
 
 __all__ = [
-    "sample_tests",
-    "fitter_tests",
     "wrapper_tests",
     "type_hints_tests",
 ]
@@ -15,8 +13,6 @@ __all__ = [
 # IMPORTS
 
 # PROJECT-SPECIFIC
-from . import test_fitter as fitter_tests
-from . import test_sample as sample_tests
 from . import test_type_hints as type_hints_tests
 from . import test_wrapper as wrapper_tests